changed menu controllers so that only a owner can change and linked to owner

+++ controllers/menu.js
@@ -2,7 +2,14 @@ const Menu = require('../models/menu')
 
 async function createdMenu(req, res) {
     try {
-        const createdItem = await Menu.create(req.body)
+        if (req.user.role !== "owner") {
+            return (res.status(403).json({ error: "Unauthorized Access"}))
+        }
+
+        const createdItem = await Menu.create({
+            ...req.body,
+            owner: req.user._id
+        })
         res.status(201).json(createdItem)
     }
     catch (error) {
@@ -12,7 +19,7 @@ async function createdMenu(req, res) {
 
 async function allMenus(req, res) {
     try {
-        const allItems = await Menu.find()
+        const allItems = await Menu.find().populate("owner", "username")
         res.status(200).json(allItems)
     }
     catch (error) {
@@ -32,6 +39,9 @@ async function showMenu(req, res) {
 
 async function updateMenu(req, res) {
     try {
+        if (req.user.role !== "owner") {
+            return (res.status(403).json({ error: "Unauthorized Access"}))
+        }
         const updatedItem = await Menu.findByIdAndUpdate(req.params.id, req.body, { new: true })
         res.status(200).json(updatedItem)
     }
@@ -42,6 +52,9 @@ async function updateMenu(req, res) {
 
 async function deleteMenu(req, res) {
     try {
+        if (req.user.role !== "owner") {
+            return (res.status(403).json({ error: "Unauthorized Access"}))
+        }
         const deletedItem = await Menu.findByIdAndDelete(req.params.id)
         res.status(200).json(deletedItem)
     }
